Include description for Deployments
In talking about this with some people, there is a bit of confusion on how to do go further (specifically deploying to Sandboxes)

I wanted to include an additional section that could be followed to make that process a bit clearer for people.

Please take a look and let me know what you think.
-Cheers.